+++ Tests/SentryTests/Integrations/Performance/UIViewController/SentryUIViewControllerPerformanceTrackerTests.swift
@@ -1,5 +1,6 @@
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
 
+import Nimble
 import ObjectiveC
 import SentryTestUtils
 import XCTest
@@ -21,6 +22,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
     let spanName = "spanName"
     let spanOperation = "spanOperation"
     let origin = "auto.ui.view_controller"
+    let frameDuration = 0.0016
     
     private class Fixture {
         
@@ -251,7 +253,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
             callbackExpectation.fulfill()
         }
         try assertSpanDuration(span: lastSpan, expectedDuration: 5)
-        try assertSpanDuration(span: transactionSpan, expectedDuration: 22)
+        try assertSpanDuration(span: transactionSpan, expectedDuration: 22 + frameDuration)
         
         wait(for: [callbackExpectation], timeout: 0)
     }
@@ -259,6 +261,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
     func testReportFullyDisplayed() {
         let sut = fixture.getSut()
         sut.enableWaitForFullDisplay = true
+        sut.enablePerformanceV2 = false
         let viewController = fixture.viewController
         let tracker = fixture.tracker
         var tracer: SentryTracer?
@@ -267,11 +270,42 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
             let spans = self.getStack(tracker)
             tracer = spans.first as? SentryTracer
         }
+        sut.viewControllerViewWillAppear(viewController) {
+            self.advanceTime(bySeconds: 0.1)
+        }
+        
+        sut.reportFullyDisplayed()
+        let expectedTTFDTimestamp = fixture.dateProvider.date()
+        reportFrame()
+        
+        let ttfdSpan = tracer?.children[1]
+        expect(ttfdSpan?.isFinished) == true
+        expect(ttfdSpan?.timestamp) == expectedTTFDTimestamp
+    }
+    
+    func testReportFullyDisplayed_WithPerformanceV2() {
+        let sut = fixture.getSut()
+        sut.enableWaitForFullDisplay = true
+        sut.enablePerformanceV2 = true
+        let viewController = fixture.viewController
+        let tracker = fixture.tracker
+        var tracer: SentryTracer?
+
+        sut.viewControllerLoadView(viewController) {
+            let spans = self.getStack(tracker)
+            tracer = spans.first as? SentryTracer
+        }
+        sut.viewControllerViewWillAppear(viewController) {
+            self.advanceTime(bySeconds: 0.1)
+        }
 
         sut.reportFullyDisplayed()
         reportFrame()
+        let expectedTTFDTimestamp = fixture.dateProvider.date()
 
-        XCTAssertTrue(tracer?.children[1].isFinished ?? false)
+        let ttfdSpan = tracer?.children[1]
+        expect(ttfdSpan?.isFinished) == true
+        expect(ttfdSpan?.timestamp) == expectedTTFDTimestamp
     }
 
     func testSecondViewController() {
@@ -598,7 +632,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
         let timestamp = try XCTUnwrap(span.timestamp)
         let startTimestamp = try XCTUnwrap(span.startTimestamp)
         let duration = timestamp.timeIntervalSince(startTimestamp)
-        XCTAssertEqual(duration, expectedDuration)
+        expect(duration).to(beCloseTo(expectedDuration, within: 0.001))
     }
     
     private func assertTrackerIsEmpty(_ tracker: SentryPerformanceTracker) {
@@ -621,6 +655,7 @@ class SentryUIViewControllerPerformanceTrackerTests: XCTestCase {
     }
 
     private func reportFrame() {
+        advanceTime(bySeconds: self.frameDuration)
         Dynamic(SentryDependencyContainer.sharedInstance().framesTracker).displayLinkCallback()
     }
 }